# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "次元链接 API",
    "description": "Dimension Link 微服务接口。成功响应为 `{ code: 0, message: 'ok', data }`；失败时 HTTP 状态码按语义返回，body 同样包一层。鉴权头：`Authorization: Bearer {accessToken}`。",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "当前服务"
    },
    {
      "url": "http://127.0.0.1:8080",
      "description": "本地开发"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "登录或注册返回的 accessToken"
      }
    },
...
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 1006
          },
          "message": {
            "type": "string",
            "example": "请先登录"
          },
          "data": {
            "nullable": true,
            "example": null
          }
        }
      },
      "UserPublic": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "u_me"
          },
          "nickname": {
            "type": "string",
            "example": "星野铃"
          },
          "handle": {
            "type": "string",
            "example": "@hoshi_suzu"
          },
          "bio": {
            "type": "string"
          },
          "signature": {
            "type": "string"
          },
          "emoji": {
            "type": "string",
            "example": "🎀"
          },
          "accentIndex": {
            "type": "integer",
            "minimum": 0,
            "maximum": 7
          },
          "followers": {
            "type": "integer"
          },
          "following": {
            "type": "integer"
          },
          "level": {
            "type": "integer"
          },
          "badges": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isFollowing": {
            "type": "boolean"
          },
          "joinedCircleIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "仅 GET /v1/me 返回"
          }
        }
      },
      "AuthPayload": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "expiresIn": {
            "type": "integer",
            "example": 7200,
            "description": "accessToken 有效期（秒）"
          },
          "user": {
            "$ref": "#/components/schemas/UserPublic"
          }
        }
      },
...
      "CircleItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "emoji": {
            "type": "string"
          },
          "desc": {
            "type": "string"
          },
          "memberCount": {
            "type": "integer"
          },
          "accentIndex": {
            "type": "integer"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "joined": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/v1/auth/register": {
      "post": {
        "tags": [
          "认证"
        ],
        "summary": "注册并登录",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "nickname",
                  "password"
                ],
                "properties": {
                  "nickname": {
                    "type": "string",
                    "example": "星野铃",
                    "description": "去空白后至少 2 个字符"
                  },
                  "handle": {
                    "type": "string",
                    "example": "suzu",
                    "description": "不要强求带 @，空则生成 user_{n}"
                  },
                  "password": {
                    "type": "string",
                    "example": "123456",
                    "description": "至少 4 位"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "注册成功即登录",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AuthPayload"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "这个 @ 已经被占用啦",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1005,
                  "message": "这个 @ 已经被占用啦",
                  "data": null
                }
              }
            }
          },
          "422": {
            "description": "昵称再可爱一点点",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1003,
                  "message": "昵称再可爱一点点",
                  "data": null
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/login": {
      "post": {
        "tags": [
          "认证"
        ],
        "summary": "登录",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "identifier",
                  "password"
                ],
                "properties": {
                  "identifier": {
                    "type": "string",
                    "example": "星野铃",
                    "description": "昵称、hoshi_suzu 或 @hoshi_suzu"
                  },
                  "password": {
                    "type": "string",
                    "example": "123456"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AuthPayload"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "通行证口令不对哦",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1002,
                  "message": "通行证口令不对哦",
                  "data": null
                }
              }
            }
          },
          "404": {
            "description": "找不到这位次元住民",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1001,
                  "message": "找不到这位次元住民",
                  "data": null
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "tags": [
          "认证"
        ],
        "summary": "刷新 token",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "refreshToken"
                ],
                "properties": {
                  "refreshToken": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "返回新的 token 对",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AuthPayload"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "请先登录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1006,
                  "message": "请先登录",
                  "data": null
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/logout": {
      "post": {
        "tags": [
          "认证"
        ],
        "summary": "登出",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "description": "可选，作废对应 refresh"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "已作废会话",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "请先登录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1006,
                  "message": "请先登录",
                  "data": null
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/circles/{circleId}": {
      "get": {
        "tags": [
          "圈子"
        ],
        "summary": "圈子详情",
        "parameters": [
          {
            "name": "circleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "c_doujin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "单个圈子",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "$ref": "#/components/schemas/CircleItem"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "圈子不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1008,
                  "message": "圈子不存在",
                  "data": null
                }
              }
            }
          }
        }
      }
    },
    "/v1/circles/{circleId}/join": {
      "post": {
        "tags": [
          "圈子"
        ],
        "summary": "加入 / 退出",
        "parameters": [
          {
            "name": "circleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "toggle 后的状态",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "joined": {
                          "type": "boolean"
                        },
                        "memberCount": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}